Changed libray to a header only type Fixed: make_room_table, copy the contents from old table_str
